Refactor datasources
